requirements.txt
future>=0.15.2
urllib3>=1.10
certifi
python-telegram-bot>=5.3.0
